drop logging, move to utils
