+++ apps/jobs/twitter-contribution/twitter_contribution/utils.py
@@ -1,3 +1,5 @@
+import logging
+import json_log_formatter
 from datetime import datetime, timedelta, timezone
 from twitter_contribution.constants import LOOKBACK_SECONDS
 
@@ -8,3 +10,15 @@ def get_lookback_window() -> datetime:
 
 def now() -> datetime:
     return datetime.now(timezone.utc)
+
+
+formatter = json_log_formatter.JSONFormatter(
+    "%(asctime)s [%(threadName)-12.12s] [%(levelname)-5.5s]  %(message)s"
+)
+
+
+def get_logger(name: str):
+    logger = logging.getLogger(name)
+    logger.setLevel(logging.INFO)
+
+    return logger
